pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=1.0.5"]

[tool.poetry]
authors = ["zakladniy"]
description = "Telegram bot for "
name = "weather_telegram_bot"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = ">=3.8.0,<3.10"
python-telegram-bot = "^13.11"
requests = "^2.27.1"
pytest = "^7.0.1"
bandit = "^1.7.4"
black = "^22.3.0"

[tool.poetry.dev-dependencies]
safety = ">=1.10.3"
coverage = ">=5.5"
nitpick = ">=0.23"
pre-commit = ">=2.6.0"
pre-commit-hooks = ">=3.2.0"
wemake-python-styleguide = ">=0.14"

# Code Formatters
autoflake = ">=1.3.1"
autopep8 = ">=1.5.4"
docformatter = ">=1.3.1"
isort = ">=4.3.21" # TODO (isort): not ^5.5.4 due to wemake-python-styleguide
language-formatters-pre-commit-hooks = ">=1.5.0"

# (Static) Code Analyzers
darglint = ">=1.5"
flake8 = ">=3.8.3"
flake8-bandit = ">=2.1"
flake8-broken-line = ">=0.2.0"
flake8-bugbear = ">=19.8"
flake8-comprehensions = ">=3.2.0" # Better list/set/dict comprehensions.
flake8-debugger = ">=3.2.0"
flake8-docstrings = ">=1.5"
flake8-eradicate = ">=0.3" # Find commented out (or so called "dead") code.
flake8-fixme = ">=1.1.0" # Check for FIXME, TODO and other temporary developer notes.
flake8-isort = ">=3.0"
flake8-logging-format = ">=0.6.0"
flake8-pytest-style = ">=1.3"
flake8-quotes = ">=2.0"
flake8-rst-docstrings = ">=0.0.12"
flake8-string-format = ">=0.2"
flake8_commas = ">=2.0"
mypy = ">=0.782"
pycodestyle = ">=2.6"
pydocstyle = ">=5.1"

# Test Suite
httpx = ">=0.16.1"
